Fix animation floating point precision issue
The maximum consecutive representable integer in a 32bit IEEE float is 16777216. This corresponds to approximately 4.66 hrs when representing milliseconds. Past this limit, integers start getting skipped with increasing frequency, causing choppy animations. A 64bit IEEE double can properly represent the full range of 32bit integers for the immediate calculation.

+++ daemon/src/engine/renderer/tr_backend.cpp
@@ -814,7 +814,7 @@ static void RB_SetGL2D()
 
 	// set time for 2D shaders
 	backEnd.refdef.time = ri.Milliseconds();
-	backEnd.refdef.floatTime = backEnd.refdef.time * 0.001f;
+	backEnd.refdef.floatTime =float(double(backEnd.refdef.time) * 0.001);
 }
 
 // used as bitfield
